+++ dpi/src/lib.rs
@@ -54,13 +54,23 @@
 //!
 //! * `serde`: Enables serialization/deserialization of certain types with [Serde](https://crates.io/crates/serde).
 //! * `mint`: Enables mint (math interoperability standard types) conversions.
+//! * `std` (enabled by default): Uses the standard library mathematical functions (normally through
+//!   your target platform's libm).
+//! * `libm`: Uses the Rust implementations of mathematical functions provided by [libm].
 //!
+//! At least one of `std` and `libm` is required to compile DPI.
+//! To use this library on a target without the standard library available, you should disable
+//! default features and enable `libm`.
 //!
 //! [points]: https://en.wikipedia.org/wiki/Point_(typography)
 //! [picas]: https://en.wikipedia.org/wiki/Pica_(typography)
 
 #![cfg_attr(docsrs, feature(doc_auto_cfg, doc_cfg_hide), doc(cfg_hide(doc, docsrs)))]
 #![forbid(unsafe_code)]
+#![no_std]
+
+#[cfg(any(feature = "std", test))]
+extern crate std;
 
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
@@ -74,32 +84,32 @@ pub trait Pixel: Copy + Into<f64> {
 
 impl Pixel for u8 {
     fn from_f64(f: f64) -> Self {
-        f.round() as u8
+        round(f) as u8
     }
 }
 impl Pixel for u16 {
     fn from_f64(f: f64) -> Self {
-        f.round() as u16
+        round(f) as u16
     }
 }
 impl Pixel for u32 {
     fn from_f64(f: f64) -> Self {
-        f.round() as u32
+        round(f) as u32
     }
 }
 impl Pixel for i8 {
     fn from_f64(f: f64) -> Self {
-        f.round() as i8
+        round(f) as i8
     }
 }
 impl Pixel for i16 {
     fn from_f64(f: f64) -> Self {
-        f.round() as i16
+        round(f) as i16
     }
 }
 impl Pixel for i32 {
     fn from_f64(f: f64) -> Self {
-        f.round() as i32
+        round(f) as i32
     }
 }
 impl Pixel for f32 {
@@ -111,6 +121,22 @@ impl Pixel for f64 {
     fn from_f64(f: f64) -> Self {
         f
     }
+}
+
+#[cfg(not(any(feature = "std", feature = "libm")))]
+compile_error!("Please select a feature to build DPI: `std`, `libm`.");
+
+#[allow(unreachable_code/* , reason = "These are intentional short-circuits" */)]
+/// Round f to the closest integer, roudning away from `0.0``
+#[inline]
+fn round(f: f64) -> f64 {
+    #[cfg(feature = "std")]
+    return f.round();
+    #[cfg(feature = "libm")]
+    return libm::round(f);
+    unreachable!(
+        "This library fails to compile if neither of the `std` or `libm` features are enabled"
+    )
 }
 
 /// Checks that the scale factor is a normal positive `f64`.
@@ -1270,20 +1296,20 @@ mod tests {
     // Eat coverage for the Debug impls et al
     #[test]
     fn ensure_attrs_do_not_panic() {
-        let _ = format!("{:?}", LogicalPosition::<u32>::default().clone());
+        let _ = std::format!("{:?}", LogicalPosition::<u32>::default().clone());
         HashSet::new().insert(LogicalPosition::<u32>::default());
 
-        let _ = format!("{:?}", PhysicalPosition::<u32>::default().clone());
+        let _ = std::format!("{:?}", PhysicalPosition::<u32>::default().clone());
         HashSet::new().insert(PhysicalPosition::<u32>::default());
 
-        let _ = format!("{:?}", LogicalSize::<u32>::default().clone());
+        let _ = std::format!("{:?}", LogicalSize::<u32>::default().clone());
         HashSet::new().insert(LogicalSize::<u32>::default());
 
-        let _ = format!("{:?}", PhysicalSize::<u32>::default().clone());
+        let _ = std::format!("{:?}", PhysicalSize::<u32>::default().clone());
         HashSet::new().insert(PhysicalSize::<u32>::default());
 
-        let _ = format!("{:?}", Size::Physical((1, 2).into()).clone());
-        let _ = format!("{:?}", Position::Physical((1, 2).into()).clone());
+        let _ = std::format!("{:?}", Size::Physical((1, 2).into()).clone());
+        let _ = std::format!("{:?}", Position::Physical((1, 2).into()).clone());
     }
 
     #[test]